globster cleanup
